+++ guillotina/db/cache/dummy.py
@@ -2,6 +2,9 @@ from guillotina import configure
 from guillotina.db.cache.base import BaseCache
 from guillotina.db.interfaces import ITransaction
 from guillotina.db.interfaces import ITransactionCache
+from typing import Any
+from typing import Dict
+from typing import List
 
 
 @configure.adapter(for_=ITransaction, provides=ITransactionCache, name="dummy")
@@ -9,17 +12,19 @@ class DummyCache(BaseCache):
     async def get(self, **kwargs):
         return None
 
-    async def set(self, value, **kwargs):
-        pass
+    async def set(
+        self, value, keyset: List[Dict[str, Any]] = None, oid=None, container=None, id=None, variant=None
+    ):
+        ...
 
     async def clear(self):
-        pass
+        ...
 
     async def invalidate(self, ob):
-        pass
+        ...
 
     async def delete(self, key):
-        pass
+        ...
 
     async def delete_all(self, keys):
-        pass
+        ...